pyproject.toml
[tool]
[tool.poetry]
name = "pdfparser"
version = "0.5.2"
homepage = "https://github.com/DarkDemiurg/pdfparser"
description = "PDF parser library."
authors = ["Dmitriy Efimov <daefimov@gmail.com>"]
readme = "README.md"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
]
packages = [
    { include = "pdfparser" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.9.*,<3.11"
click = "8.1.8"
"pdfminer.six" = "^20220524"
PyPDF2 = "^2.12.1"
tabula-py = "^2.10.0"
pdftables-api = {git = "https://github.com/pdftables/python-pdftables-api.git"}
PyMuPDF = "^1.24.11"
pdftotext = "^3.0.0"

black  = { version = "^25.1.0", optional = true}
isort  = { version = "^6.0.0", optional = true}
flake8  = { version = "^7.1.1", optional = true}
flake8-docstrings = { version = "^1.6.0", optional = true }
mypy = {version = "^1.14.1", optional = true}
pytest  = { version = "^8.3.4", optional = true}
pytest-cov  = { version = "^6.0.0", optional = true}
tox  = { version = "^4.24.1", optional = true}
virtualenv  = { version = "^20.16.2", optional = true}
pip  = { version = "^25.0.0", optional = true}
mkdocs  = { version = "^1.3.1", optional = true}
mkdocs-autorefs = {version = "^1.3.0", optional = true}
mkdocs-include-markdown-plugin  = { version = "^7.1.2", optional = true}
mkdocs-material  = { version = "^9.6.2", optional = true}
mkdocs-material-extensions  = { version = "^1.0.3", optional = true}
mkdocstrings  = {extras = ["python"], version = "^0.28.0"}
twine  = { version = "^6.1.0", optional = true}
pre-commit = {version = "^4.1.0", optional = true}
toml = {version = "^0.10.2", optional = true}
bump2version = {version = "^1.0.1", optional = true}
requests-mock = {version = "^1.9.3", optional = true}

[tool.poetry.extras]
test = [
    "pytest",
    "black",
    "isort",
    "mypy",
    "flake8",
    "flake8-docstrings",
    "pytest-cov",
    "PyPDF2",
    "pdfminer.six",
    "tabula-py",
    "pdftables-api",
    "requests-mock",
    "pdftotext"
    ]

dev = ["tox", "pre-commit", "virtualenv", "pip", "twine", "toml", "bump2version"]

doc = [
    "mkdocs",
    "mkdocs-include-markdown-plugin",
    "mkdocs-material",
    "mkdocstrings",
    "mkdocs-material-extension",
    "mkdocs-autorefs"
    ]

[tool.poetry.scripts]
pdfparser = 'pdfparser.cli:main'

[[tool.poetry.source]]
name = "pypi-public"
url = "https://pypi.org/simple/"

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py39', 'py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
# you can skip files as below
#skip_glob = docs/conf.py

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
